Adding a "force" parameter in files to fix #242.  Currently only used when replacing directories with links, but should probably be used in other places.
git-svn-id: https://reductivelabs.com/svn/puppet/trunk@1538 980ebf18-57e1-0310-9a29-db15c13687c0

+++ lib/puppet/type/pfile/target.rb
@@ -28,7 +28,12 @@ module Puppet
 
                 case stat.ftype
                 when "directory":
-                    FileUtils.rmtree(@parent[:path])
+                    if @parent[:force] == :true
+                        FileUtils.rmtree(@parent[:path])
+                    else
+                        notice "Not replacing directory with link; use 'force' to override"
+                        return :nochange
+                    end
                 else
                     File.unlink(@parent[:path])
                 end
